Rename package "tasks" to "task"
