+++ src/main/java/task/Deadline.java
@@ -1,15 +1,15 @@
-package tasks;
+package task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Event extends Task {
+public class Deadline extends Task {
     private String taskName;
     private String time;
-    private String taskSymbol = "E";
+    private String taskSymbol = "D";
 
-    public Event(String taskName, String time) {
+    public Deadline(String taskName, String time) {
         this.taskName = taskName;
         this.time = time;
     }
@@ -32,13 +32,12 @@ public class Event extends Task {
     @Override
     public String toString() {
         String statusIcon = getStatusIcon();
-        return String.format("[%s][%s] %s (at: %s)", taskSymbol, statusIcon, taskName, timeToString());
+        return String.format("[%s][%s] %s (by: %s)", taskSymbol, statusIcon, taskName, timeToString());
     }
 
     @Override
     public String toStorageFormat() {
         return String.format("%s/%s/%s/%s", taskSymbol, isCompleted(), taskName, time);
     }
+
 }
-
-
